client_api/requirements.txt
fastapi
uvicorn
sqlalchemy==1.4.40
psycopg2
python-dotenv